Renamed AtheleteSeedTimeView to AthleteSeedTimeView.
